tests/requirements.txt
pytest
requests
openai
langchain
langchain-openai
pydantic
python-dotenv
aiohttp
